+++ lib/index.ts
@@ -8,7 +8,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-var requireIndex = require('requireindex')
+import requireIndex from 'requireindex'
 
 //------------------------------------------------------------------------------
 // Plugin Definition